refactor: directory function into new dirs module
