+++ .bumpversion.cfg
@@ -1,5 +1,5 @@
 [bumpversion]
-current_version = 0.8.5-dev6
+current_version = 1.0.0-dev0
 commit = False
 allow_dirty = True
 tag = False
