+++ .vscode/settings.json
@@ -10,9 +10,13 @@
     "ansible.python.interpreterPath": "/home/jim/.cache/pypoetry/virtualenvs/wtforglib-C4UfSdlt-py3.10/bin/python",
     "cSpell.words": [
         "DMESSAGE",
+        "dpath",
+        "dspec",
         "dtrace",
         "EMESSAGE",
         "euid",
+        "filenotfound",
+        "fstats",
         "geteuid",
         "IMESSAGE",
         "isdebug",
@@ -23,6 +27,7 @@
         "isverbose",
         "itrace",
         "tmplwrtr",
+        "topdown",
         "VMESSAGE",
         "WMESSAGE"
     ],
