+++ pyproject.toml
@@ -1,7 +1,7 @@
 [tool.poetry]
 name = "wtforglib"
 description = "A set of utility functions I use in my various python projects."
-version = "0.8.5-dev6"
+version = "1.0.0-dev0"
 license = "MIT"
 
 authors = [
